requirements.txt
PySide6==6.6.1
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.0
Flask-CORS==4.0.0
openpyxl==3.1.2